second chance for 0x02 project

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -24,8 +24,8 @@
  *
  */void fibon(int n)
 {
-	unsigned int term1 = 1, term2 = 2;
-	unsigned int i, next;
+	unsigned long long term1 = 1, term2 = 2;
+	unsigned long long i, next;
 
 	printf("1, 2, ");
